# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Noteora Smart Data Analytics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  active: "active",
  draft: "draft",
  archived: "archived",
} as const;

export interface Project {
  id: number;
  userId: string;
  name: string;
  description?: string | null;
  status: ProjectStatus;
  datasetCount: number;
  chartCount: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateProjectBodyStatus =
  (typeof CreateProjectBodyStatus)[keyof typeof CreateProjectBodyStatus];

export const CreateProjectBodyStatus = {
  active: "active",
  draft: "draft",
  archived: "archived",
} as const;

export interface CreateProjectBody {
  name: string;
  description?: string | null;
  status?: CreateProjectBodyStatus;
}

export type UpdateProjectBodyStatus =
  (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus];

export const UpdateProjectBodyStatus = {
  active: "active",
  draft: "draft",
  archived: "archived",
} as const;

export interface UpdateProjectBody {
  name?: string;
  description?: string | null;
  status?: UpdateProjectBodyStatus;
}

export type DatasetSourceType =
  (typeof DatasetSourceType)[keyof typeof DatasetSourceType];

export const DatasetSourceType = {
  csv: "csv",
  excel: "excel",
  manual: "manual",
  api: "api",
} as const;

export interface Dataset {
  id: number;
  projectId: number;
  userId: string;
  name: string;
  sourceType: DatasetSourceType;
  columns: string[];
  rowCount: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateDatasetBodySourceType =
  (typeof CreateDatasetBodySourceType)[keyof typeof CreateDatasetBodySourceType];

export const CreateDatasetBodySourceType = {
  csv: "csv",
  excel: "excel",
  manual: "manual",
  api: "api",
} as const;

export interface CreateDatasetBody {
  projectId: number;
  name: string;
  sourceType: CreateDatasetBodySourceType;
  columns: string[];
}

export type DatasetRowsResponseRowsItem = { [key: string]: unknown };

export interface DatasetRowsResponse {
  rows: DatasetRowsResponseRowsItem[];
  total: number;
  page: number;
  limit: number;
}

export type AddDatasetRowsBodyRowsItem = { [key: string]: unknown };

export interface AddDatasetRowsBody {
  rows: AddDatasetRowsBodyRowsItem[];
}

export type ChartChartType =
  (typeof ChartChartType)[keyof typeof ChartChartType];

export const ChartChartType = {
  line: "line",
  bar: "bar",
  pie: "pie",
  area: "area",
  donut: "donut",
  table: "table",
  kpi: "kpi",
  comparison: "comparison",
} as const;

export type ChartConfig = { [key: string]: unknown } | null;

export interface Chart {
  id: number;
  projectId: number;
  datasetId?: number | null;
  userId: string;
  title: string;
  chartType: ChartChartType;
  xAxis?: string | null;
  yAxis?: string | null;
  config?: ChartConfig;
  pinnedToDashboard: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateChartBodyChartType =
  (typeof CreateChartBodyChartType)[keyof typeof CreateChartBodyChartType];

export const CreateChartBodyChartType = {
  line: "line",
  bar: "bar",
  pie: "pie",
  area: "area",
  donut: "donut",
  table: "table",
  kpi: "kpi",
  comparison: "comparison",
} as const;

export type CreateChartBodyConfig = { [key: string]: unknown } | null;

export interface CreateChartBody {
  projectId: number;
  datasetId?: number | null;
  title: string;
  chartType: CreateChartBodyChartType;
  xAxis?: string | null;
  yAxis?: string | null;
  config?: CreateChartBodyConfig;
  pinnedToDashboard?: boolean;
}

export type UpdateChartBodyChartType =
  (typeof UpdateChartBodyChartType)[keyof typeof UpdateChartBodyChartType];

export const UpdateChartBodyChartType = {
  line: "line",
  bar: "bar",
  pie: "pie",
  area: "area",
  donut: "donut",
  table: "table",
  kpi: "kpi",
  comparison: "comparison",
} as const;

export type UpdateChartBodyConfig = { [key: string]: unknown } | null;

export interface UpdateChartBody {
  title?: string;
  chartType?: UpdateChartBodyChartType;
  xAxis?: string | null;
  yAxis?: string | null;
  config?: UpdateChartBodyConfig;
  pinnedToDashboard?: boolean;
}

export interface Report {
  id: number;
  projectId: number;
  userId: string;
  title: string;
  summary?: string | null;
  chartIds: number[];
  dateRangeStart?: string | null;
  dateRangeEnd?: string | null;
  createdAt: string;
}

export interface CreateReportBody {
  projectId: number;
  title: string;
  summary?: string | null;
  chartIds: number[];
  dateRangeStart?: string | null;
  dateRangeEnd?: string | null;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  info: "info",
  warning: "warning",
  success: "success",
  error: "error",
} as const;

export interface Notification {
  id: number;
  userId: string;
  title: string;
  message: string;
  type: NotificationType;
  isRead: boolean;
  createdAt: string;
}

export interface DashboardSummary {
  totalProjects: number;
  totalDatasets: number;
  totalCharts: number;
  totalReports: number;
  totalRows: number;
  activeProjects: number;
  unreadNotifications: number;
  recentProjectNames: string[];
}

export type InsightType = (typeof InsightType)[keyof typeof InsightType];

export const InsightType = {
  trend: "trend",
  warning: "warning",
  opportunity: "opportunity",
  summary: "summary",
} as const;

export type InsightSeverity =
  (typeof InsightSeverity)[keyof typeof InsightSeverity];

export const InsightSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Insight {
  id: string;
  type: InsightType;
  title: string;
  message: string;
  severity: InsightSeverity;
  createdAt: string;
}

export type ActivityItemEntityType =
  (typeof ActivityItemEntityType)[keyof typeof ActivityItemEntityType];

export const ActivityItemEntityType = {
  project: "project",
  dataset: "dataset",
  chart: "chart",
  report: "report",
} as const;

export interface ActivityItem {
  id: number;
  action: string;
  entityType: ActivityItemEntityType;
  entityName: string;
  createdAt: string;
}

export interface TrendPoint {
  month: string;
  projects: number;
  datasets: number;
  charts: number;
  reports: number;
}

export type GetDatasetRowsParams = {
  page?: number;
  limit?: number;
};

export type AddDatasetRows201 = {
  inserted: number;
};

export type MarkAllNotificationsRead200 = {
  updated: number;
};