# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Noteora Smart Data Analytics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/datasets:
    get:
      operationId: listProjectDatasets
      tags: [datasets]
      summary: List datasets for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dataset"
  /datasets:
    get:
      operationId: listDatasets
      tags: [datasets]
      summary: List all datasets for authenticated user
      responses:
        "200":
          description: List of datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dataset"
    post:
      operationId: createDataset
      tags: [datasets]
      summary: Create a dataset (manual entry or CSV upload metadata)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDatasetBody"
      responses:
        "201":
          description: Created dataset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
  /datasets/{id}:
    get:
      operationId: getDataset
      tags: [datasets]
      summary: Get a dataset by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dataset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
    delete:
      operationId: deleteDataset
      tags: [datasets]
      summary: Delete a dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /datasets/{id}/rows:
    get:
      operationId: getDatasetRows
      tags: [datasets]
      summary: Get rows for a dataset (paginated)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Dataset rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetRowsResponse"
    post:
      operationId: addDatasetRows
      tags: [datasets]
      summary: Add rows to a dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDatasetRowsBody"
      responses:
        "201":
          description: Rows added
          content:
            application/json:
              schema:
                type: object
                properties:
                  inserted:
                    type: integer
                required:
                  - inserted
  /charts:
    get:
      operationId: listCharts
      tags: [charts]
      summary: List all charts for authenticated user
      responses:
        "200":
          description: List of charts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chart"
    post:
      operationId: createChart
      tags: [charts]
      summary: Create a chart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChartBody"
      responses:
        "201":
          description: Created chart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chart"
  /charts/{id}:
    get:
      operationId: getChart
      tags: [charts]
      summary: Get a chart by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chart"
    put:
      operationId: updateChart
      tags: [charts]
      summary: Update a chart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChartBody"
      responses:
        "200":
          description: Updated chart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chart"
    delete:
      operationId: deleteChart
      tags: [charts]
      summary: Delete a chart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Dataset:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        userId:
          type: string
        name:
          type: string
        sourceType:
          type: string
          enum: [csv, excel, manual, api]
        columns:
          type: array
          items:
            type: string
        rowCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - userId
        - name
        - sourceType
        - columns
        - rowCount
        - createdAt
        - updatedAt
    CreateDatasetBody:
      type: object
      properties:
        projectId:
          type: integer
        name:
          type: string
        sourceType:
          type: string
          enum: [csv, excel, manual, api]
        columns:
          type: array
          items:
            type: string
      required:
        - projectId
        - name
        - sourceType
        - columns
    DatasetRowsResponse:
      type: object
      properties:
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - rows
        - total
        - page
        - limit
    AddDatasetRowsBody:
      type: object
      properties:
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
      required:
        - rows
    Chart:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        datasetId:
          type: integer
          nullable: true
        userId:
          type: string
        title:
          type: string
        chartType:
          type: string
          enum: [line, bar, pie, area, donut, table, kpi, comparison]
        xAxis:
          type: string
          nullable: true
        yAxis:
          type: string
          nullable: true
        config:
          type: object
          additionalProperties: true
          nullable: true
        pinnedToDashboard:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - userId
        - title
        - chartType
        - pinnedToDashboard
        - createdAt
        - updatedAt
    CreateChartBody:
      type: object
      properties:
        projectId:
          type: integer
        datasetId:
          type: integer
          nullable: true
        title:
          type: string
        chartType:
          type: string
          enum: [line, bar, pie, area, donut, table, kpi, comparison]
        xAxis:
          type: string
          nullable: true
        yAxis:
          type: string
          nullable: true
        config:
          type: object
          additionalProperties: true
          nullable: true
        pinnedToDashboard:
          type: boolean
          default: false
      required:
        - projectId
        - title
        - chartType
    UpdateChartBody:
      type: object
      properties:
        title:
          type: string
        chartType:
          type: string
          enum: [line, bar, pie, area, donut, table, kpi, comparison]
        xAxis:
          type: string
          nullable: true
        yAxis:
          type: string
          nullable: true
        config:
          type: object
          additionalProperties: true
          nullable: true
        pinnedToDashboard:
          type: boolean
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, warning, success, error]
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - title
        - message
        - type
        - isRead
        - createdAt
...